Refactor: code cleanup and garbage collection
