+++ apps/mobile/src/App.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useMemo, useRef, useState } from 'react';
+import React, { useMemo, useCallback } from 'react';
 import { BottomNav } from './components/BottomNav';
 import { Welcome } from './components/screens/Welcome';
 import { Login } from './components/screens/Login';
@@ -22,204 +22,65 @@ import { TermsAndConditions } from './components/screens/TermsAndConditions';
 import { PrivacyPolicy } from './components/screens/PrivacyPolicy';
 import { TitoliOttenuti } from './components/screens/TitoliOttenuti';
 import { GameStateProvider, useGameState } from './state/store';
-import { isSupabaseConfigured, supabase } from './lib/supabase';
-import { getPermission, notify } from './lib/notifications';
+import { isSupabaseConfigured } from './lib/supabase';
+import { openInMaps, openEventsMap } from './lib/navigation-utils';
+import { uploadProfileImage } from './services/storage';
 
-
-type Screen =
-  | 'welcome'
-  | 'login'
-  | 'signup'
-  | 'install'
-  | 'role-selection'
-  | 'home'
-  | 'turni'
-  | 'leaderboard'
-  | 'qr-scanner'
-  | 'event-confirmation'
-  | 'event-details'
-  | 'attivita'
-  | 'activity-detail'
-  | 'profilo'
-  | 'account-settings'
-  | 'support'
-  | 'change-password'
-  | 'carriera'
-  | 'terms'
-  | 'privacy'
-  | 'titoli-ottenuti';
-
-type Tab = 'home' | 'turni' | 'leaderboard' | 'attivita' | 'profilo';
-
-type LegalReturnScreen = Exclude<Screen, 'terms' | 'privacy'>;
-
-const NAV_STATE_KEY = 'tdp-mobile-ui-nav';
-const NAV_STATE_VERSION = 1 as const;
-
-type PersistedNavState = {
-  version: typeof NAV_STATE_VERSION;
-  screen: Screen;
-  activeTab: Tab;
-  legalReturnScreen: LegalReturnScreen;
-  isPasswordRecovery: boolean;
-  scannedEventId: string;
-  selectedActivityId: string;
-};
-
-const VALID_SCREENS = new Set<Screen>([
-  'welcome',
-  'login',
-  'signup',
-  'install',
-  'role-selection',
-  'home',
-  'turni',
-  'leaderboard',
-  'qr-scanner',
-  'event-confirmation',
-  'event-details',
-  'attivita',
-  'activity-detail',
-  'profilo',
-  'account-settings',
-  'support',
-  'change-password',
-  'carriera',
-  'terms',
-  'privacy',
-  'titoli-ottenuti',
-]);
-
-const VALID_TABS = new Set<Tab>(['home', 'turni', 'leaderboard', 'attivita', 'profilo']);
-
-const VALID_LEGAL_RETURN_SCREENS = new Set<LegalReturnScreen>([
-  'welcome',
-  'login',
-  'signup',
-  'role-selection',
-  'home',
-  'turni',
-  'qr-scanner',
-  'event-confirmation',
-  'attivita',
-  'activity-detail',
-  'profilo',
-  'account-settings',
-  'support',
-  'change-password',
-  'carriera',
-  'titoli-ottenuti',
-]);
-
-function readNavState(): PersistedNavState | null {
-  if (typeof window === 'undefined') return null;
-  try {
-    const raw = window.localStorage.getItem(NAV_STATE_KEY);
-    if (!raw) return null;
-    const parsed = JSON.parse(raw) as Partial<PersistedNavState>;
-    if (parsed.version !== NAV_STATE_VERSION) return null;
-    if (!parsed.screen || !VALID_SCREENS.has(parsed.screen)) return null;
-    if (!parsed.activeTab || !VALID_TABS.has(parsed.activeTab)) return null;
-    if (!parsed.legalReturnScreen || !VALID_LEGAL_RETURN_SCREENS.has(parsed.legalReturnScreen)) return null;
-
-    return {
-      version: NAV_STATE_VERSION,
-      screen: parsed.screen,
-      activeTab: parsed.activeTab,
-      legalReturnScreen: parsed.legalReturnScreen,
-      isPasswordRecovery: Boolean(parsed.isPasswordRecovery),
-      scannedEventId: typeof parsed.scannedEventId === 'string' ? parsed.scannedEventId : '',
-      selectedActivityId: typeof parsed.selectedActivityId === 'string' ? parsed.selectedActivityId : '',
-    };
-  } catch {
-    return null;
-  }
-}
-
-function writeNavState(state: PersistedNavState) {
-  if (typeof window === 'undefined') return;
-  try {
-    window.localStorage.setItem(NAV_STATE_KEY, JSON.stringify(state));
-  } catch {
-    // ignore quota/security errors
-  }
-}
-
-function getScreenToPersist(screen: Screen, activeTab: Tab): Screen {
-  if (screen === 'install') {
-    return activeTab === 'home' ? 'home' : 'turni';
-  }
-  if (screen === 'qr-scanner') {
-    return activeTab === 'home' ? 'home' : 'turni';
-  }
-  return screen;
-}
+// Types and Hooks
+import { Screen, LegalReturnScreen } from './types/navigation';
+import { useNavigation } from './hooks/useNavigation';
+import { useAuth } from './hooks/useAuth';
+import { useNotifications } from './hooks/useNotifications';
+import { useQrLanding } from './hooks/useQrLanding';
 
 function AppShell() {
   const {
-    authUserId,
-    authReady,
-    hasHydratedRemote,
-    state,
-    roles,
-    events,
-    activities,
-    turnStats,
-    statsLoading,
-    theatreReputation,
-    theatreReputationLoading,
-    badges,
-    followedEvents,
-    followedEventsLoading,
-    followEvent,
-    unfollowEvent,
-    isEventFollowed,
-    markBadgesSeen,
-    updateProfile,
-    registerTurn,
-    completeActivity,
-    resetProgress,
-    changePassword,
-    sendPasswordResetEmail,
+    authUserId, authReady, hasHydratedRemote, state, roles, events,
+    activities, turnStats, statsLoading, theatreReputation,
+    theatreReputationLoading, badges, followedEvents, followedEventsLoading,
+    followEvent, unfollowEvent, isEventFollowed, markBadgesSeen,
+    updateProfile, registerTurn, completeActivity, resetProgress,
+    changePassword, sendPasswordResetEmail
   } = useGameState();
 
-  const persistedNavState = useMemo(() => {
-    const persisted = readNavState();
-    if (!persisted) return null;
-    let nextScreen = persisted.screen;
-    if (nextScreen === 'activity-detail' && !persisted.selectedActivityId) {
-      nextScreen = 'attivita';
-    }
-    if (nextScreen === 'event-confirmation' && !persisted.scannedEventId) {
-      nextScreen = 'turni';
-    }
-    nextScreen = getScreenToPersist(nextScreen, persisted.activeTab);
-    return { ...persisted, screen: nextScreen };
-  }, []);
+  // Navigation Hook
+  const {
+    currentScreen, setCurrentScreen, activeTab, handleTabChange,
+    legalReturnScreen, setLegalReturnScreen, isPasswordRecovery, setIsPasswordRecovery,
+    scannedEventId, setScannedEventId, selectedActivityId, setSelectedActivityId,
+  } = useNavigation(events);
 
-  const [currentScreen, setCurrentScreen] = useState<Screen>(() => persistedNavState?.screen ?? 'welcome');
-  const [legalReturnScreen, setLegalReturnScreen] = useState<LegalReturnScreen>(() => persistedNavState?.legalReturnScreen ?? 'welcome');
-  const [isPasswordRecovery, setIsPasswordRecovery] = useState(() => persistedNavState?.isPasswordRecovery ?? false);
-  const [activeTab, setActiveTab] = useState<Tab>(() => persistedNavState?.activeTab ?? 'home');
-  const [scannedEventId, setScannedEventId] = useState<string>(() => persistedNavState?.scannedEventId ?? events[0]?.id ?? '');
-  const [selectedActivityId, setSelectedActivityId] = useState<string>(() => persistedNavState?.selectedActivityId ?? '');
-  const [authError, setAuthError] = useState<string | null>(null);
-  const currentScreenRef = useRef(currentScreen);
-  const lastNotifiedBadgeId = useRef<string | null>(null);
-  const lastNotifiedEventId = useRef<string | null>(null);
-  const hasHandledQrLanding = useRef(false);
-
-  const upcomingEvent = useMemo(() => followedEvents[0], [followedEvents]);
-  const unlockedBadges = useMemo(() => badges.filter((badge) => badge.unlocked), [badges]);
-  const newBadges = useMemo(() => unlockedBadges.filter((badge) => !badge.seenAt), [unlockedBadges]);
-  const theatreReputationForProfile = useMemo(
-    () => theatreReputation.map((entry) => ({ name: entry.theatre, reputation: entry.reputation })),
-    [theatreReputation]
+  // Auth Hook
+  const {
+    authError, setAuthError, handleLogin, handleSignup, handleLogout
+  } = useAuth(
+    state.profile,
+    updateProfile,
+    (screen, isRecovery) => {
+      if (isRecovery !== undefined) setIsPasswordRecovery(isRecovery);
+      if (screen === 'home') {
+        const shouldAutoHome = currentScreen === 'welcome' || currentScreen === 'login';
+        if (shouldAutoHome) setCurrentScreen('home');
+      } else {
+        setCurrentScreen(screen);
+      }
+    },
+    () => {
+      setIsPasswordRecovery(false);
+      setCurrentScreen('welcome');
+      handleTabChange('home');
+    }
   );
+
+  // Helpers derived from state
+  const upcomingEvent = useMemo(() => followedEvents[0], [followedEvents]);
+  const unlockedBadges = useMemo(() => badges.filter(b => b.unlocked), [badges]);
+  const newBadges = useMemo(() => unlockedBadges.filter(b => !b.seenAt), [unlockedBadges]);
   const newestNewBadge = useMemo(() => {
     if (!newBadges.length) return null;
     return [...newBadges].sort((a, b) => (b.unlockedAt ?? 0) - (a.unlockedAt ?? 0))[0];
   }, [newBadges]);
+
   const hasValidEmail = Boolean(state.profile.email && state.profile.email.includes('@'));
   const isAuthValid = useMemo(() => {
     if (!isSupabaseConfigured) return true;
@@ -227,645 +88,66 @@ function AppShell() {
     return Boolean(authUserId && hasHydratedRemote && hasValidEmail);
   }, [authReady, authUserId, hasHydratedRemote, hasValidEmail]);
 
-  useEffect(() => {
-    if (!events.length) {
-      if (scannedEventId) {
-        setScannedEventId('');
-      }
-      return;
-    }
-    if (!scannedEventId || !events.some((event) => event.id === scannedEventId)) {
-      setScannedEventId(events[0].id);
-    }
-  }, [events, scannedEventId]);
+  // Notification and QR Landing Hooks
+  useNotifications(upcomingEvent, newestNewBadge ?? undefined);
+  useQrLanding(authReady, isAuthValid, (target) => setCurrentScreen(target));
 
-  useEffect(() => {
-    currentScreenRef.current = currentScreen;
-  }, [currentScreen]);
-
-  useEffect(() => {
-    if (!newestNewBadge) return;
-    if (lastNotifiedBadgeId.current === newestNewBadge.id) return;
-    if (getPermission() !== 'granted') return;
-
-    notify('Nuovo badge sbloccato', {
-      body: newestNewBadge.title,
-      tag: `badge-${newestNewBadge.id}`,
-    });
-    lastNotifiedBadgeId.current = newestNewBadge.id;
-  }, [newestNewBadge]);
-
-  useEffect(() => {
-    if (!upcomingEvent) return;
-    if (lastNotifiedEventId.current === upcomingEvent.id) return;
-    if (getPermission() !== 'granted') return;
-
-    notify('Nuovo evento in agenda', {
-      body: `${upcomingEvent.name} · ${upcomingEvent.date} ${upcomingEvent.time}`,
-      tag: `event-${upcomingEvent.id}`,
-    });
-    lastNotifiedEventId.current = upcomingEvent.id;
-  }, [upcomingEvent]);
-
-  useEffect(() => {
-    writeNavState({
-      version: NAV_STATE_VERSION,
-      screen: getScreenToPersist(currentScreen, activeTab),
-      activeTab,
-      legalReturnScreen,
-      isPasswordRecovery,
-      scannedEventId,
-      selectedActivityId,
-    });
-  }, [activeTab, currentScreen, isPasswordRecovery, legalReturnScreen, scannedEventId, selectedActivityId]);
-
-  useEffect(() => {
-    if (typeof window === 'undefined') return;
-    if (hasHandledQrLanding.current) return;
-    const search = new URLSearchParams(window.location.search);
-    if (search.get('from') !== 'qr') return;
-    if (!authReady) return;
-    hasHandledQrLanding.current = true;
-    if (!isAuthValid) {
-      setCurrentScreen('welcome');
-    } else {
-      setCurrentScreen('install');
-    }
-    try {
-      window.history.replaceState({}, '', window.location.pathname);
-    } catch {
-      // ignore
-    }
-  }, [authReady, isAuthValid]);
-
-  const handleStart = () => setCurrentScreen('signup');
-
-  const handleLogin = async (email: string, password: string) => {
-    setAuthError(null);
-    if (!isSupabaseConfigured || !supabase) {
-      updateProfile({ email });
-      setCurrentScreen('home');
-      return;
-    }
-
-    const { data, error } = await supabase.auth.signInWithPassword({ email, password });
-    if (error) {
-      setAuthError(error.message);
-      return;
-    }
-
-    const displayName = data.user?.user_metadata?.name ?? state.profile.name;
-    updateProfile({ name: displayName, email });
-    setCurrentScreen('home');
-  };
-
-  const handleSignup = async (name: string, email: string, password: string) => {
-    setAuthError(null);
-    if (!isSupabaseConfigured || !supabase) {
-      updateProfile({ name, email });
-      setCurrentScreen('role-selection');
-      return;
-    }
-
-    const { data, error } = await supabase.auth.signUp({
-      email,
-      password,
-      options: { data: { name } },
-    });
-    if (error) {
-      setAuthError(error.message);
-      return;
-    }
-
-    const displayName = data.user?.user_metadata?.name ?? name;
-    updateProfile({ name: displayName, email });
-    setCurrentScreen('role-selection');
-  };
-
-  const handleRoleComplete = (roleId: string) => {
-    updateProfile({ roleId: roleId as typeof state.profile.roleId });
-    setCurrentScreen('home');
-  };
-
-  const handleTabChange = (tab: Tab) => {
-    setActiveTab(tab);
-    switch (tab) {
-      case 'home':
-        setCurrentScreen('home');
-        break;
-      case 'turni':
-        setCurrentScreen('turni');
-        break;
-      case 'leaderboard':
-        setCurrentScreen('leaderboard');
-        break;
-      case 'attivita':
-        setCurrentScreen('attivita');
-        break;
-      case 'profilo':
-        setCurrentScreen('profilo');
-        break;
-    }
-  };
-
+  // Handler Actions
   const handleQRScanAttempt = (code: string) => {
-    if (!authReady) {
-      return { ok: false as const, error: 'Sto verificando la sessione, riprova tra poco.' };
-    }
-    if (!isAuthValid) {
-      setCurrentScreen('welcome');
-      return { ok: false as const, error: 'Effettua il login per continuare.' };
-    }
-    if (!events.length) {
-      return { ok: false as const, error: 'Nessun evento disponibile al momento.' };
-    }
-    const resolved = events.find((event) => code.toLowerCase().includes(event.id.toLowerCase()));
-    if (!resolved) {
-      return { ok: false as const, error: 'Questo QR non sembra appartenere a un biglietto ATCL.' };
-    }
-
+    if (!authReady) return { ok: false as const, error: 'Verifica sessione...' };
+    if (!isAuthValid) { setCurrentScreen('welcome'); return { ok: false as const, error: 'Login richiesto.' }; }
+    const resolved = events.find(e => code.toLowerCase().includes(e.id.toLowerCase()));
+    if (!resolved) return { ok: false as const, error: 'QR non valido.' };
     setScannedEventId(resolved.id);
     setCurrentScreen('event-confirmation');
     return { ok: true as const };
   };
 
   const handleEventConfirm = () => {
-    const resolvedRoleId = state.profile.roleId;
-    const record = registerTurn(scannedEventId, resolvedRoleId);
-    if (record) {
-      setCurrentScreen('home');
-      setActiveTab('home');
+    if (registerTurn(scannedEventId, state.profile.roleId)) {
+      handleTabChange('home');
     }
   };
 
-  const handleStartActivity = (activityId: string) => {
-    setSelectedActivityId(activityId);
-    setCurrentScreen('activity-detail');
+  const handleUploadImage = async (file: File) => {
+    if (!authUserId) return;
+    const url = await uploadProfileImage(authUserId, file);
+    updateProfile({ profileImage: url });
   };
 
-  const handleActivityStart = () => {
-    if (selectedActivityId) {
-      completeActivity(selectedActivityId);
-    }
-    setCurrentScreen('attivita');
-    setActiveTab('attivita');
-  };
-
-  const handleViewCarriera = () => {
-    setCurrentScreen('carriera');
-  };
-
-  const handleViewEventDetails = () => {
-    if (!upcomingEvent) return;
-    setScannedEventId(upcomingEvent.id);
-    setCurrentScreen('event-details');
-  };
-
-  const handleNavigateToEvent = (event?: { theatre: string } | null) => {
-    const targetEvent = event ?? upcomingEvent;
-    if (!targetEvent) return;
-    const destination = encodeURIComponent(targetEvent.theatre);
-    if (typeof window === 'undefined') return;
-    const isAppleDevice = /iPad|iPhone|iPod/i.test(navigator.userAgent);
-    if (isAppleDevice) {
-      window.location.href = `maps://?q=${destination}`;
-      return;
-    }
-    if (/Android/i.test(navigator.userAgent)) {
-      window.location.href = `geo:0,0?q=${destination}`;
-      return;
-    }
-    window.open(`https://www.google.com/maps/dir/?api=1&destination=${destination}`, '_blank', 'noopener');
-  };
-
-  const handleViewEventsMap = () => {
-    if (typeof window === 'undefined') return;
-    const uniqueTheatres = Array.from(new Set(events.map((event) => event.theatre).filter(Boolean)));
-    if (!uniqueTheatres.length) return;
-    const origin = encodeURIComponent('My Location');
-    const destination = encodeURIComponent(uniqueTheatres[uniqueTheatres.length - 1]);
-    const waypoints = uniqueTheatres
-      .slice(0, -1)
-      .map((value) => encodeURIComponent(value))
-      .join('|');
-    const url = `https://www.google.com/maps/dir/?api=1&origin=${origin}&destination=${destination}${
-      waypoints ? `&waypoints=${waypoints}` : ''
-    }`;
-    window.location.href = url;
-  };
-
-  const handleToggleFollow = (eventId: string) => {
-    if (isEventFollowed(eventId)) {
-      void unfollowEvent(eventId);
-    } else {
-      void followEvent(eventId);
-    }
-  };
-
-  const handleViewTitoli = () => {
-    setCurrentScreen('titoli-ottenuti');
-  };
-
-  const handleViewAccountSettings = () => {
-    setCurrentScreen('account-settings');
-  };
-
-  const handleViewSupport = () => {
-    setCurrentScreen('support');
-  };
-
-  const handleOpenQRScanner = () => {
-    if (!authReady) {
-      return;
-    }
-    if (!isAuthValid) {
-      setCurrentScreen('welcome');
-      return;
-    }
-    setCurrentScreen('qr-scanner');
-  };
-
-  const openTerms = (from: LegalReturnScreen) => {
+  const openLegal = (screen: 'terms' | 'privacy', from: LegalReturnScreen) => {
     setLegalReturnScreen(from);
-    setCurrentScreen('terms');
+    setCurrentScreen(screen);
   };
-
-  const openPrivacy = (from: LegalReturnScreen) => {
-    setLegalReturnScreen(from);
-    setCurrentScreen('privacy');
-  };
-
-  const handleLogout = () => {
-    if (supabase) {
-      supabase.auth.signOut();
-    }
-    setIsPasswordRecovery(false);
-    setCurrentScreen('welcome');
-    setActiveTab('home');
-  };
-
-  useEffect(() => {
-    if (!isSupabaseConfigured) return;
-    if (!authReady || !authUserId) return;
-    if (hasHydratedRemote && hasValidEmail) return;
-    if (typeof window === 'undefined') return;
-
-    const timeoutId = window.setTimeout(() => {
-      if (!hasHydratedRemote || !hasValidEmail) {
-        handleLogout();
-      }
-    }, 4000);
-
-    return () => {
-      window.clearTimeout(timeoutId);
-    };
-  }, [authReady, authUserId, handleLogout, hasHydratedRemote, hasValidEmail]);
-
-  const handleUploadProfileImage = async (file: File) => {
-    if (!supabase || !authUserId) {
-      throw new Error('Supabase non configurato o utente non autenticato');
-    }
-
-    // Create a unique filename
-    const fileExt = file.name.split('.').pop();
-    const fileName = `${authUserId}/profile.${fileExt}`;
-
-    // Upload to Supabase storage
-    const { data, error } = await supabase.storage
-      .from('profile-images')
-      .upload(fileName, file, {
-        cacheControl: '3600',
-        upsert: true,
-      });
-
-    if (error) {
-      throw error;
-    }
-
-    // Get public URL
-    const { data: urlData } = supabase.storage
-      .from('profile-images')
-      .getPublicUrl(fileName);
-
-    if (!urlData.publicUrl) {
-      throw new Error('Impossibile ottenere l\'URL pubblico');
-    }
-
-    // Update profile with the image URL
-    updateProfile({ profileImage: urlData.publicUrl });
-  };
-
-  const handleResetProgress = async () => {
-    if (typeof window === 'undefined') return;
-    try {
-      await resetProgress();
-      setActiveTab('home');
-      setCurrentScreen('role-selection');
-    } catch (error) {
-      const message = error instanceof Error ? error.message : 'Errore sconosciuto';
-      window.alert(`Impossibile resettare i progressi: ${message}`);
-    }
-  };
-
-  const handleCloseChangePassword = () => {
-    setIsPasswordRecovery(false);
-    setCurrentScreen(isPasswordRecovery ? 'home' : 'account-settings');
-  };
-
-  useEffect(() => {
-    setAuthError(null);
-  }, [currentScreen]);
-
-  useEffect(() => {
-    if (!supabase) return;
-    let mounted = true;
-
-    supabase.auth.getSession().then(({ data, error }) => {
-      if (!mounted || error) return;
-      if (data.session?.user) {
-        const user = data.session.user;
-        const displayName = user.user_metadata?.name ?? state.profile.name;
-        updateProfile({ name: displayName, email: user.email ?? state.profile.email });
-        const shouldAutoHome =
-          currentScreenRef.current === 'welcome' || currentScreenRef.current === 'login';
-        if (shouldAutoHome) {
-          setCurrentScreen('home');
-        }
-      }
-    });
-
-    const { data: authListener } = supabase.auth.onAuthStateChange((event, session) => {
-      if (!mounted) return;
-      if (event === 'SIGNED_OUT') {
-        setIsPasswordRecovery(false);
-        setCurrentScreen('welcome');
-        setActiveTab('home');
-        return;
-      }
-      if (event === 'PASSWORD_RECOVERY') {
-        if (session?.user) {
-          const displayName = session.user.user_metadata?.name ?? state.profile.name;
-          updateProfile({ name: displayName, email: session.user.email ?? state.profile.email });
-        }
-        setIsPasswordRecovery(true);
-        setCurrentScreen('change-password');
-        return;
-      }
-      if (session?.user) {
-        const displayName = session.user.user_metadata?.name ?? state.profile.name;
-        updateProfile({ name: displayName, email: session.user.email ?? state.profile.email });
-        const shouldAutoHome =
-          currentScreenRef.current === 'welcome' || currentScreenRef.current === 'login';
-        if (shouldAutoHome) {
-          setCurrentScreen('home');
-        }
-      }
-    });
-
-    return () => {
-      mounted = false;
-      authListener?.subscription.unsubscribe();
-    };
-  }, [updateProfile, state.profile.email, state.profile.name]);
-
-  const selectedEvent = events.find((event) => event.id === scannedEventId) ?? undefined;
-  const currentActivity = activities.find((item) => item.id === selectedActivityId);
 
   const renderScreen = () => {
+    const selectedEvent = events.find(e => e.id === scannedEventId);
+    const selectedRole = roles.find(r => r.id === state.profile.roleId);
+    const currentActivity = activities.find(a => a.id === selectedActivityId);
+
     switch (currentScreen) {
-      case 'welcome':
-        return <Welcome onStart={handleStart} onLogin={() => setCurrentScreen('login')} />;
-
-      case 'login':
-        return (
-          <Login
-            onBack={() => setCurrentScreen('welcome')}
-            onLogin={handleLogin}
-            onSignup={() => setCurrentScreen('signup')}
-            onForgotPassword={() => undefined}
-            errorMessage={authError}
-          />
-        );
-
-      case 'signup':
-        return (
-          <Signup
-            onBack={() => setCurrentScreen('welcome')}
-            onSignup={handleSignup}
-            onLogin={() => setCurrentScreen('login')}
-            onViewTerms={() => openTerms('signup')}
-            onViewPrivacy={() => openPrivacy('signup')}
-            errorMessage={authError}
-          />
-        );
-
-      case 'install':
-        return (
-          <InstallApp
-            onContinue={() => {
-              if (!isAuthValid) {
-                setCurrentScreen('welcome');
-                return;
-              }
-              setCurrentScreen(state.profile.roleId ? 'home' : 'welcome');
-            }}
-          />
-        );
-
-      case 'role-selection':
-        return <RoleSelection roles={roles} onComplete={(role) => handleRoleComplete(role.id)} />;
-
-      case 'home':
-        return (
-          <Home
-            userName={state.profile.name}
-            userRole={roles.find((role) => role.id === state.profile.roleId)?.name ?? 'Ruolo'}
-            level={state.profile.level}
-            xp={state.profile.xp}
-            xpToNextLevel={state.profile.xpToNextLevel}
-            reputation={state.profile.reputation}
-            onScanQR={handleOpenQRScanner}
-            onViewActivities={() => handleTabChange('attivita')}
-            onViewTurni={() => handleTabChange('turni')}
-            onViewEventDetails={handleViewEventDetails}
-            onNavigateToEvent={() => handleNavigateToEvent(upcomingEvent)}
-            upcomingEvent={upcomingEvent}
-            totalTurns={turnStats.totalTurns}
-            turnsThisMonth={turnStats.turnsThisMonth}
-            uniqueTheatres={turnStats.uniqueTheatres}
-            activitiesCount={activities.length}
-            eventLoading={followedEventsLoading}
-            statsLoading={statsLoading}
-            newBadgesCount={newBadges.length}
-            newBadgeTitle={newestNewBadge?.title ?? undefined}
-            onDismissBadgeNotification={markBadgesSeen}
-          />
-        );
-
-      case 'turni':
-        return (
-          <TurniATCL
-            events={events}
-            isEventFollowed={isEventFollowed}
-            onToggleFollow={handleToggleFollow}
-            onViewMap={handleViewEventsMap}
-            onViewEvent={(eventId) => {
-              setScannedEventId(eventId);
-              setCurrentScreen('event-details');
-            }}
-            onScanQR={handleOpenQRScanner}
-          />
-        );
-
-      case 'leaderboard':
-        return <Leaderboard />;
-
-      case 'qr-scanner':
-        return (
-          <QRScanner
-            onClose={() => {
-              const previousTab = activeTab === 'home' ? 'home' : 'turni';
-              handleTabChange(previousTab);
-            }}
-            onScan={handleQRScanAttempt}
-          />
-        );
-
-      case 'event-confirmation': {
-        const selectedRole = roles.find((role) => role.id === state.profile.roleId);
-        return (
-          <EventConfirmation
-            event={selectedEvent}
-            role={selectedRole}
-            onConfirm={handleEventConfirm}
-            onCancel={() => setCurrentScreen('turni')}
-          />
-        );
-      }
-
-      case 'event-details':
-        return (
-          <EventDetails
-            event={selectedEvent}
-            onBack={() => setCurrentScreen('home')}
-            onNavigate={() => handleNavigateToEvent(selectedEvent)}
-          />
-        );
-
-      case 'attivita':
-        return <Attivita activities={activities} onStartActivity={handleStartActivity} />;
-
-      case 'activity-detail':
-        return (
-          currentActivity && (
-            <ActivityDetail
-              activity={currentActivity}
-              onStart={handleActivityStart}
-              onClose={() => setCurrentScreen('attivita')}
-            />
-          )
-        );
-
-      case 'profilo':
-        return (
-          <Profilo
-            userName={state.profile.name}
-            userRole={roles.find((role) => role.id === state.profile.roleId)?.name ?? 'Ruolo'}
-            level={state.profile.level}
-            xp={state.profile.xp}
-            xpTotal={state.profile.xpTotal}
-            xpSulCampo={state.profile.xpField}
-            reputationGlobal={state.profile.reputation}
-            theatreReputation={theatreReputationForProfile}
-            theatreReputationLoading={theatreReputationLoading}
-            badgesUnlockedCount={unlockedBadges.length}
-            newBadgesCount={newBadges.length}
-            profileImage={state.profile.profileImage}
-            onViewCarriera={handleViewCarriera}
-            onViewTitoli={handleViewTitoli}
-            onSettings={handleViewAccountSettings}
-            onLogout={handleLogout}
-            onUploadProfileImage={handleUploadProfileImage}
-          />
-        );
-
-      case 'account-settings':
-        return (
-          <AccountSettings
-            userName={state.profile.name}
-            email={state.profile.email}
-            onBack={() => setCurrentScreen('profilo')}
-            onViewTerms={() => openTerms('account-settings')}
-            onViewPrivacy={() => openPrivacy('account-settings')}
-            onViewSupport={handleViewSupport}
-            onChangePassword={() => {
-              setIsPasswordRecovery(false);
-              setCurrentScreen('change-password');
-            }}
-            onResetProgress={handleResetProgress}
-            onLogout={handleLogout}
-          />
-        );
-
-      case 'support':
-        return (
-          <SupportChat
-            userName={state.profile.name}
-            onBack={() => setCurrentScreen('account-settings')}
-          />
-        );
-
-      case 'change-password':
-        return (
-          <ChangePassword
-            email={state.profile.email}
-            mode={isPasswordRecovery ? 'recovery' : 'change'}
-            onBack={handleCloseChangePassword}
-            onChangePassword={(currentPassword, newPassword) => changePassword(newPassword, currentPassword)}
-            onSendResetEmail={() => sendPasswordResetEmail(state.profile.email)}
-          />
-        );
-
-      case 'carriera': {
-        const selectedRole = roles.find((role) => role.id === state.profile.roleId);
-        const fallbackRoleStats = { presence: 0, precision: 0, leadership: 0, creativity: 0 };
-        return (
-          <Carriera
-            userRole={selectedRole?.name ?? 'Ruolo'}
-            roleId={state.profile.roleId}
-            roleStats={selectedRole?.stats ?? fallbackRoleStats}
-            turnStats={turnStats}
-            badges={badges}
-            turns={state.turns}
-            roles={roles}
-            level={state.profile.level}
-            xp={state.profile.xp}
-            xpToNextLevel={state.profile.xpToNextLevel}
-            xpTotal={state.profile.xpTotal}
-            xpSulCampo={state.profile.xpField}
-            reputationGlobal={state.profile.reputation}
-            onBack={() => setCurrentScreen('profilo')}
-          />
-        );
-      }
-
-      case 'terms':
-        return <TermsAndConditions onBack={() => setCurrentScreen(legalReturnScreen)} />;
-
-      case 'privacy':
-        return <PrivacyPolicy onBack={() => setCurrentScreen(legalReturnScreen)} />;
-
-      case 'titoli-ottenuti':
-        return (
-          <TitoliOttenuti
-            badges={unlockedBadges}
-            onBack={() => setCurrentScreen('profilo')}
-            onViewed={markBadgesSeen}
-          />
-        );
-
-      default:
-        return null;
+      case 'welcome': return <Welcome onStart={() => setCurrentScreen('signup')} onLogin={() => setCurrentScreen('login')} />;
+      case 'login': return <Login onBack={() => setCurrentScreen('welcome')} onLogin={handleLogin} onSignup={() => setCurrentScreen('signup')} onForgotPassword={() => { }} errorMessage={authError} />;
+      case 'signup': return <Signup onBack={() => setCurrentScreen('welcome')} onSignup={handleSignup} onLogin={() => setCurrentScreen('login')} onViewTerms={() => openLegal('terms', 'signup')} onViewPrivacy={() => openLegal('privacy', 'signup')} errorMessage={authError} />;
+      case 'install': return <InstallApp onContinue={() => setCurrentScreen(state.profile.roleId ? 'home' : 'welcome')} />;
+      case 'role-selection': return <RoleSelection roles={roles} onComplete={(role) => { updateProfile({ roleId: role.id as any }); setCurrentScreen('home'); }} />;
+      case 'home': return <Home userName={state.profile.name} userRole={selectedRole?.name ?? 'Ruolo'} level={state.profile.level} xp={state.profile.xp} xpToNextLevel={state.profile.xpToNextLevel} reputation={state.profile.reputation} onScanQR={() => setCurrentScreen('qr-scanner')} onViewActivities={() => handleTabChange('attivita')} onViewTurni={() => handleTabChange('turni')} onViewEventDetails={() => { setScannedEventId(upcomingEvent?.id ?? ''); setCurrentScreen('event-details'); }} onNavigateToEvent={() => openInMaps(upcomingEvent?.theatre ?? '')} upcomingEvent={upcomingEvent} totalTurns={turnStats.totalTurns} turnsThisMonth={turnStats.turnsThisMonth} uniqueTheatres={turnStats.uniqueTheatres} activitiesCount={activities.length} eventLoading={followedEventsLoading} statsLoading={statsLoading} newBadgesCount={newBadges.length} newBadgeTitle={newestNewBadge?.title} onDismissBadgeNotification={markBadgesSeen} />;
+      case 'turni': return <TurniATCL events={events} isEventFollowed={isEventFollowed} onToggleFollow={(id) => isEventFollowed(id) ? unfollowEvent(id) : followEvent(id)} onViewMap={() => openEventsMap(events.map(e => e.theatre))} onViewEvent={(id) => { setScannedEventId(id); setCurrentScreen('event-details'); }} onScanQR={() => setCurrentScreen('qr-scanner')} />;
+      case 'leaderboard': return <Leaderboard />;
+      case 'qr-scanner': return <QRScanner onClose={() => handleTabChange(activeTab === 'home' ? 'home' : 'turni')} onScan={handleQRScanAttempt} />;
+      case 'event-confirmation': return <EventConfirmation event={selectedEvent} role={selectedRole} onConfirm={handleEventConfirm} onCancel={() => setCurrentScreen('turni')} />;
+      case 'event-details': return <EventDetails event={selectedEvent} onBack={() => setCurrentScreen('home')} onNavigate={() => openInMaps(selectedEvent?.theatre ?? '')} />;
+      case 'attivita': return <Attivita activities={activities} onStartActivity={(id) => { setSelectedActivityId(id); setCurrentScreen('activity-detail'); }} />;
+      case 'activity-detail': return currentActivity && <ActivityDetail activity={currentActivity} onStart={() => { completeActivity(selectedActivityId); handleTabChange('attivita'); }} onClose={() => setCurrentScreen('attivita')} />;
+      case 'profilo': return <Profilo userName={state.profile.name} userRole={selectedRole?.name ?? 'Ruolo'} level={state.profile.level} xp={state.profile.xp} xpTotal={state.profile.xpTotal} xpSulCampo={state.profile.xpField} reputationGlobal={state.profile.reputation} theatreReputation={theatreReputation} theatreReputationLoading={theatreReputationLoading} badgesUnlockedCount={unlockedBadges.length} newBadgesCount={newBadges.length} profileImage={state.profile.profileImage} onViewCarriera={() => setCurrentScreen('carriera')} onViewTitoli={() => setCurrentScreen('titoli-ottenuti')} onSettings={() => setCurrentScreen('account-settings')} onLogout={handleLogout} onUploadProfileImage={handleUploadImage} />;
+      case 'account-settings': return <AccountSettings userName={state.profile.name} email={state.profile.email} onBack={() => setCurrentScreen('profilo')} onViewTerms={() => openLegal('terms', 'account-settings')} onViewPrivacy={() => openLegal('privacy', 'account-settings')} onViewSupport={() => setCurrentScreen('support')} onChangePassword={() => { setIsPasswordRecovery(false); setCurrentScreen('change-password'); }} onResetProgress={async () => { await resetProgress(); handleTabChange('home'); setCurrentScreen('role-selection'); }} onLogout={handleLogout} />;
+      case 'support': return <SupportChat userName={state.profile.name} onBack={() => setCurrentScreen('account-settings')} />;
+      case 'change-password': return <ChangePassword email={state.profile.email} mode={isPasswordRecovery ? 'recovery' : 'change'} onBack={() => { setIsPasswordRecovery(false); setCurrentScreen(isPasswordRecovery ? 'home' : 'account-settings'); }} onChangePassword={changePassword} onSendResetEmail={() => sendPasswordResetEmail(state.profile.email)} />;
+      case 'carriera': return <Carriera userRole={selectedRole?.name ?? 'Ruolo'} roleId={state.profile.roleId} roleStats={selectedRole?.stats ?? {}} turnStats={turnStats} badges={badges} turns={state.turns} roles={roles} level={state.profile.level} xp={state.profile.xp} xpToNextLevel={state.profile.xpToNextLevel} xpTotal={state.profile.xpTotal} xpSulCampo={state.profile.xpField} reputationGlobal={state.profile.reputation} onBack={() => setCurrentScreen('profilo')} />;
+      case 'terms': return <TermsAndConditions onBack={() => setCurrentScreen(legalReturnScreen)} />;
+      case 'privacy': return <PrivacyPolicy onBack={() => setCurrentScreen(legalReturnScreen)} />;
+      case 'titoli-ottenuti': return <TitoliOttenuti badges={unlockedBadges} onBack={() => setCurrentScreen('profilo')} onViewed={markBadgesSeen} />;
+      default: return null;
     }
   };
